Add test for mutually recursive remote functions. (#5349)
